% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPtree.R
\name{DPTreePrior}
\alias{DPTreePrior}
\title{Generating the standard D-P Tree prior}
\usage{
DPTreePrior(m = 4, z = 1)
}
\arguments{
\item{m}{A positive integer. The finite approximation level for D-P tree. Default m=4.}

\item{z}{A positive number. On i-th level, the hyperparameter for D-P tree prior is
\eqn{z\times i^2}. Default z=1.}
}
\value{
A list.
\item{a}{An array containing the hyperparameters of D-P trees.}
}
\description{
\code{DPTreePrior} returns
a standard D-P Tree prior based on specified hyperparameters.
}
\examples{
DPTreePrior(m=6, z=1)
}
\references{
\insertRef{DPtree}{DPtree}
}
