% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPtree.R
\name{RealizeDPTree}
\alias{RealizeDPTree}
\title{Sampling a realized distribution from the D-P Tree.}
\usage{
RealizeDPTree(prior)
}
\arguments{
\item{prior}{A list. Should be in same format as returned from  \code{DPTreePrior}.}
}
\value{
An array of dimension  \eqn{2^m} by \eqn{2^m} by m. m is the approximation level.
 Realized Z's for all partitions at each level.
Three dimensions reprensent two marginals, and the level respectively.
}
\description{
\code{RealizeDPTree} returns
a realized (copula) distribtuion sampled from the input D-P Tree.
}
\examples{
RealizeDPTree(DPTreePrior(m=2, z=1))
}
\references{
\insertRef{DPtree}{DPtree}
}
